requirements.txt
aiogram==3.4.1
asyncio==3.4.3
pillow==10.2.0
